Created a logger
Before I was using the root logger because that's the example I had.
But with uvicorn, I think it was subsuming the debug messages. At this
point in time I wanted to have debug messages so I could try and make sure
of what was happening before I started writing tests.