Moved current_games to within all the if statements for the webhook endpoint
I had having to repeat stuff here, but (since you're not supposed to
repeat because it means a change in one place may not get propagated
everywhere). But I wonder if sending the message to Matrix is what allowed a
delay in writing to the dictionary that allows for a double-post.

However, now that I think about it, perhaps I could use a boolean to
dictate whether to send a message and save it for the end. That would result
in even less repetition.

First I need to see whether I'm on the right track and this fixes things.